Create face_recog_LED .py
The face_recog_LED.py code implements both face recognition and face registration functions. Specifically, when a face is detected, the system compares the current face features with the registered face feature library. If the similarity score is higher than the preset threshold, the face is considered recognized. If the face has not been registered, a white box is drawn around the detected face. If the face has already been registered, a green box is drawn, and each registered person will display a different expression (such as a heart, smile, angry face, etc.) on the micro:bit LED screen (up to 10 people).
When the BOOT_KEY button is pressed, the feature of the currently detected face will be added to the feature library for future recognition.